Completed: Project basic level 👍

+++ pages/api/products/[id].js
@@ -4,8 +4,10 @@ import dbConnect from 'utils/mongo'
 export default async function handler(req, res) {
 	const {
 		method,
+		cookies,
 		query: { id },
 	} = req
+	const token = cookies.token
 	await dbConnect()
 
 	switch (method) {
@@ -19,6 +21,9 @@ export default async function handler(req, res) {
 			break
 
 		case 'POST':
+			if (!token || token !== process.env.TOKEN) {
+				return res.status(401).json('Not authenticated!')
+			}
 			try {
 				const product = await Product.create(req.body)
 				res.status(200).json(product)
@@ -28,6 +33,9 @@ export default async function handler(req, res) {
 			break
 
 		case 'PUT':
+			if (!token || token !== process.env.TOKEN) {
+				return res.status(401).json('Not authenticated!')
+			}
 			try {
 			} catch (err) {
 				res.status(500).json(err)
@@ -35,6 +43,9 @@ export default async function handler(req, res) {
 			break
 
 		case 'DELETE':
+			if (!token || token !== process.env.TOKEN) {
+				return res.status(401).json('Not authenticated!')
+			}
 			try {
 				await Product.findByIdAndDelete(id)
 				res.status(200).json('Product has been deleted!')
